Do not check ssr state in handleSsrStopRequest
The snapshotter is not set to active state until an initial delta
snapshot or a fullsnapshot has been taken. If a leadership election
takes place while the snapshot is being taken and the current leader
loses that election, then handleSsrRequest does not send on the
ssrStopCh. This in turn causes ssr.snapshotEventHandler to never return
and snapshots would continue to be taken.