Do not check SSR State when Snapshotter should be stopped (#680) (#750)
* Specify send/recv

Allows type checking of channel interactions

* Close instead of sending on channel

I think that closing the channel makes it clearer that this channels purpose has been fulfilled and it no longer needs to be considered. All the places that use the channel are recieving all values from the channel so they will continue to function in the same way

* Do not check ssr state in handleSsrStopRequest

The snapshotter is not set to active state until an initial delta
snapshot or a fullsnapshot has been taken. If a leadership election
takes place while the snapshot is being taken and the current leader
loses that election, then handleSsrRequest does not send on the
ssrStopCh. This in turn causes ssr.snapshotEventHandler to never return
and snapshots would continue to be taken.

Co-authored-by: Alex Vest <avestuk@gmail.com>